refactor(native-filters): refactor code of native filters
