+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterBar.tsx
@@ -16,36 +16,21 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import {
-  QueryFormData,
-  styled,
-  SuperChart,
-  t,
-  ExtraFormData,
-} from '@superset-ui/core';
-import React, { useState, useEffect, useMemo, useRef } from 'react';
+import { styled, t, ExtraFormData } from '@superset-ui/core';
+import React, { useState, useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import cx from 'classnames';
 import Button from 'src/components/Button';
 import Icon from 'src/components/Icon';
-import { getChartDataRequest } from 'src/chart/chartAction';
-import { areObjectsEqual } from 'src/reduxUtils';
-import Loading from 'src/components/Loading';
-import BasicErrorAlert from 'src/components/ErrorMessage/BasicErrorAlert';
+import { CurrentFilterState } from 'src/dashboard/reducers/types';
 import FilterConfigurationLink from './FilterConfigurationLink';
 import {
-  useCascadingFilters,
   useFilterConfiguration,
   useFilters,
-  useFilterState,
   useSetExtraFormData,
 } from './state';
-import { Filter, CascadeFilter, CurrentFilterState } from './types';
-import {
-  buildCascadeFiltersTree,
-  getFormData,
-  mapParentFiltersToChildren,
-} from './utils';
+import { Filter } from '../types';
+import { buildCascadeFiltersTree, mapParentFiltersToChildren } from './utils';
 import CascadePopover from './CascadePopover';
 
 const barWidth = `250px`;
@@ -55,10 +40,6 @@ const BarWrapper = styled.div`
   &.open {
     width: ${barWidth}; // arbitrary...
   }
-`;
-
-const FilterItem = styled.div`
-  padding-bottom: 10px;
 `;
 
 const Bar = styled.div`
@@ -157,229 +138,11 @@ const FilterControls = styled.div`
   padding: ${({ theme }) => theme.gridUnit * 4}px;
 `;
 
-const StyledCascadeChildrenList = styled.ul`
-  list-style-type: none;
-  & > * {
-    list-style-type: none;
-  }
-`;
-
-const StyledFilterControlTitle = styled.h4`
-  width: 100%;
-  font-size: ${({ theme }) => theme.typography.sizes.s}px;
-  color: ${({ theme }) => theme.colors.grayscale.dark1};
-  margin: 0;
-  overflow-wrap: break-word;
-`;
-
-const StyledFilterControlTitleBox = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: ${({ theme }) => theme.gridUnit}px;
-`;
-
-const StyledFilterControlContainer = styled.div`
-  width: 100%;
-`;
-
-const StyledFilterControlBox = styled.div`
-  display: flex;
-`;
-
-const StyledCaretIcon = styled(Icon)`
-  margin-top: ${({ theme }) => -theme.gridUnit}px;
-`;
-
-const StyledLoadingBox = styled.div`
-  position: relative;
-  height: ${({ theme }) => theme.gridUnit * 8}px;
-  margin-bottom: ${({ theme }) => theme.gridUnit * 6}px;
-`;
-
-interface FilterProps {
-  filter: Filter;
-  icon?: React.ReactElement;
-  directPathToChild?: string[];
-  onFilterSelectionChange: (
-    filter: Filter,
-    extraFormData: ExtraFormData,
-    currentState: CurrentFilterState,
-  ) => void;
-}
-
 interface FiltersBarProps {
   filtersOpen: boolean;
   toggleFiltersBar: any;
   directPathToChild?: string[];
 }
-
-const FilterValue: React.FC<FilterProps> = ({
-  filter,
-  directPathToChild,
-  onFilterSelectionChange,
-}) => {
-  const {
-    id,
-    allowsMultipleValues,
-    inverseSelection,
-    targets,
-    defaultValue,
-    filterType,
-  } = filter;
-  const cascadingFilters = useCascadingFilters(id);
-  const filterState = useFilterState(id);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [state, setState] = useState([]);
-  const [error, setError] = useState<boolean>(false);
-  const [formData, setFormData] = useState<Partial<QueryFormData>>({});
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [target] = targets;
-  const { datasetId = 18, column } = target;
-  const { name: groupby } = column;
-  const currentValue = filterState.currentState?.value;
-  useEffect(() => {
-    const newFormData = getFormData({
-      datasetId,
-      cascadingFilters,
-      groupby,
-      allowsMultipleValues,
-      defaultValue,
-      currentValue,
-      inverseSelection,
-    });
-    if (!areObjectsEqual(formData || {}, newFormData)) {
-      setFormData(newFormData);
-      getChartDataRequest({
-        formData: newFormData,
-        force: false,
-        requestParams: { dashboardId: 0 },
-      })
-        .then(response => {
-          setState(response.result);
-          setError(false);
-          setLoading(false);
-        })
-        .catch(() => {
-          setError(true);
-          setLoading(false);
-        });
-    }
-  }, [cascadingFilters, datasetId, groupby, defaultValue, currentValue]);
-
-  useEffect(() => {
-    if (directPathToChild?.[0] === filter.id) {
-      // wait for Cascade Popover to open
-      const timeout = setTimeout(() => {
-        inputRef?.current?.focus();
-      }, 200);
-      return () => clearTimeout(timeout);
-    }
-    return undefined;
-  }, [inputRef, directPathToChild, filter.id]);
-
-  const setExtraFormData = ({
-    extraFormData,
-    currentState,
-  }: {
-    extraFormData: ExtraFormData;
-    currentState: CurrentFilterState;
-  }) => onFilterSelectionChange(filter, extraFormData, currentState);
-
-  if (loading) {
-    return (
-      <StyledLoadingBox>
-        <Loading />
-      </StyledLoadingBox>
-    );
-  }
-
-  if (error) {
-    return (
-      <BasicErrorAlert
-        title={t('Cannot load filter')}
-        body={t('Check configuration')}
-        level="error"
-      />
-    );
-  }
-
-  return (
-    <FilterItem>
-      <SuperChart
-        height={20}
-        width={220}
-        formData={formData}
-        queriesData={state}
-        chartType={filterType}
-        // @ts-ignore (update superset-ui)
-        hooks={{ setExtraFormData }}
-      />
-    </FilterItem>
-  );
-};
-
-export const FilterControl: React.FC<FilterProps> = ({
-  filter,
-  icon,
-  onFilterSelectionChange,
-  directPathToChild,
-}) => {
-  const { name = '<undefined>' } = filter;
-  return (
-    <StyledFilterControlContainer>
-      <StyledFilterControlTitleBox>
-        <StyledFilterControlTitle>{name}</StyledFilterControlTitle>
-        <div>{icon}</div>
-      </StyledFilterControlTitleBox>
-      <FilterValue
-        filter={filter}
-        directPathToChild={directPathToChild}
-        onFilterSelectionChange={onFilterSelectionChange}
-      />
-    </StyledFilterControlContainer>
-  );
-};
-
-interface CascadeFilterControlProps {
-  filter: CascadeFilter;
-  directPathToChild?: string[];
-  onFilterSelectionChange: (
-    filter: Filter,
-    extraFormData: ExtraFormData,
-    currentState: CurrentFilterState,
-  ) => void;
-}
-
-export const CascadeFilterControl: React.FC<CascadeFilterControlProps> = ({
-  filter,
-  directPathToChild,
-  onFilterSelectionChange,
-}) => (
-  <>
-    <StyledFilterControlBox>
-      <StyledCaretIcon name="caret-down" />
-      <FilterControl
-        filter={filter}
-        directPathToChild={directPathToChild}
-        onFilterSelectionChange={onFilterSelectionChange}
-      />
-    </StyledFilterControlBox>
-
-    <StyledCascadeChildrenList>
-      {filter.cascadeChildren?.map(childFilter => (
-        <li key={childFilter.id}>
-          <CascadeFilterControl
-            filter={childFilter}
-            directPathToChild={directPathToChild}
-            onFilterSelectionChange={onFilterSelectionChange}
-          />
-        </li>
-      ))}
-    </StyledCascadeChildrenList>
-  </>
-);
 
 const FilterBar: React.FC<FiltersBarProps> = ({
   filtersOpen,
