+++ superset-frontend/src/dashboard/components/nativeFilters/FilterConfigModal/FilterScope.tsx
