+++ superset-frontend/src/dashboard/components/nativeFilters/FilterConfigModal/FiltersList.tsx
