+++ superset-frontend/src/dashboard/reducers/types.ts
@@ -18,6 +18,8 @@
  */
 
 import componentTypes from 'src/dashboard/util/componentTypes';
+import { ExtraFormData, JsonObject } from '@superset-ui/core';
+import { Filter } from '../components/nativeFilters/types';
 
 export enum Scoping {
   all,
@@ -43,6 +45,11 @@ export type RootState = {
 /** State of dashboardLayout in redux */
 export type Layout = { [key: string]: LayoutItem };
 
+/** State of nativeFilters currentState */
+export type CurrentFilterState = JsonObject & {
+  value: any;
+};
+
 /** State of charts in redux */
 export type Charts = { [key: number]: Chart };
 
@@ -62,5 +69,21 @@ export type LayoutItem = {
     text?: string;
     uuid: string;
     width: number;
+  };
+};
+
+/** Current state of the filter, stored in `nativeFilters` in redux */
+export type FilterState = {
+  id: string; // ties this filter state to the config object
+  extraFormData?: ExtraFormData;
+  currentState?: CurrentFilterState;
+};
+
+export type NativeFiltersState = {
+  filters: {
+    [filterId: string]: Filter;
+  };
+  filtersState: {
+    [filterId: string]: FilterState;
   };
 };
